+++ test/schemas/test_parametrization.py
@@ -140,7 +140,7 @@ def test_(request, case):
 
 
 def test_recursive_dereference(testdir):
-    # When a given parameter contains a JSON reference, that reference an object with another reference"
+    # When a given parameter contains a JSON reference, that reference an object with another reference
     testdir.make_test(
         """
 @schema.parametrize()
@@ -160,6 +160,74 @@ def test_(request, case):
                 "properties": {"id": {"$ref": "#/definitions/SimpleIntRef"}},
             },
             "SimpleIntRef": {"type": "integer"},
+        },
+    )
+    # Then it should be correctly resolved and used in the generated case
+    result = testdir.runpytest("-v", "-s")
+    result.assert_outcomes(passed=1)
+    result.stdout.re_match_lines([r"Hypothesis calls: 1"])
+
+
+def test_inner_dereference(testdir):
+    # When a given parameter contains a JSON reference inside a property of an object
+    testdir.make_test(
+        """
+@schema.parametrize()
+@settings(max_examples=1)
+def test_(request, case):
+    request.config.HYPOTHESIS_CASES += 1
+    assert case.path == "/v1/users"
+    assert case.method == "GET"
+    assert_int(case.body["id"])
+""",
+        **as_param(
+            {
+                "schema": {
+                    "type": "object",
+                    "required": ["id"],
+                    "properties": {"id": {"$ref": "#/definitions/SimpleIntRef"}},
+                },
+                "in": "body",
+                "name": "object",
+                "required": True,
+            }
+        ),
+        definitions={"SimpleIntRef": {"type": "integer"}},
+    )
+    # Then it should be correctly resolved and used in the generated case
+    result = testdir.runpytest("-v", "-s")
+    result.assert_outcomes(passed=1)
+    result.stdout.re_match_lines([r"Hypothesis calls: 1"])
+
+
+def test_inner_dereference_with_lists(testdir):
+    # When a given parameter contains a JSON reference inside a list in `allOf`
+    testdir.make_test(
+        """
+@schema.parametrize()
+@settings(max_examples=1)
+def test_(request, case):
+    request.config.HYPOTHESIS_CASES += 1
+    assert case.path == "/v1/users"
+    assert case.method == "GET"
+    assert_int(case.body["id"]["a"])
+    assert_str(case.body["id"]["b"])
+""",
+        **as_param(
+            {
+                "schema": {
+                    "type": "object",
+                    "required": ["id"],
+                    "properties": {"id": {"allOf": [{"$ref": "#/definitions/A"}, {"$ref": "#/definitions/B"}]}},
+                },
+                "in": "body",
+                "name": "object",
+                "required": True,
+            }
+        ),
+        definitions={
+            "A": {"type": "object", "required": ["a"], "properties": {"a": {"type": "integer"}}},
+            "B": {"type": "object", "required": ["b"], "properties": {"b": {"type": "string"}}},
         },
     )
     # Then it should be correctly resolved and used in the generated case