+++ test/utils.py
@@ -7,4 +7,4 @@ def get_schema_path(schema_name):
     return os.path.join(HERE, "data", schema_name)
 
 
-SIMPLE_PATH = get_schema_path("simple.yaml")
+SIMPLE_PATH = get_schema_path("simple_swagger.yaml")